+++ Domain/Interfaces/IUserRepository.cs
@@ -2,5 +2,5 @@ using Domain.Entities;
 
 namespace Domain.Interfaces;
 public interface IUserRepository: IGenericRepository<User>{
-    Task<User?> GetByIdAsync(long id);
+    Task<User?> GetByIdAsync(int id);
 }
